Moved the cache key to the utils module. Renamed the dao method.
